rename: 폴더 이름 컨벤션 적용
