+++ src/app/mypage/shuttle/components/reservation-card/HandyBadge.tsx
