+++ src/app/mypage/shuttle/components/reservation-card/ShuttleBusBadge.tsx
