+++ src/app/mypage/shuttle/components/reservation-card/hooks/useTextAndStyle.ts
